Several minor fixes
