+++ Homework/TagCloudContainerTests/TagCloudContainerTests.csproj
@@ -57,7 +57,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="WordsGetterShould.cs" />
+    <Compile Include="TextParserShould.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
