--- Homework/TagsCloudContainer/IPaintConfig.cs
@@ -1,13 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Drawing;
-
-namespace TagsCloudContainer
-{
-    public interface IPaintConfig
-    {
-        List<Pen> WordsColors { get;}
-        Font WordsFont { get;}
-
-        Size ImageSize { get; }
-    }
-}
