+++ Homework/TagsCloudContainer/TagsCloudContainer.csproj
@@ -44,11 +44,13 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CustomizingPaintCofig.cs" />
-    <Compile Include="IPaintConfig.cs" />
+    <Compile Include="CloudPainter.cs" />
+    <Compile Include="TextParsers\ITextParser.cs" />
+    <Compile Include="PaintConfigs\CustomizedPaintCofig.cs" />
+    <Compile Include="PaintConfigs\IPaintConfig.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="WordsGetter.cs" />
+    <Compile Include="TextParsers\TextParser.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
@@ -62,5 +64,6 @@
   <ItemGroup>
     <Content Include="words.txt" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
