+++ Homework/TagsCloudContainer/TextParsers/ITextParser.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace TagsCloudContainer
+{
+    public interface ITextParser
+    {
+        IReadOnlyDictionary<string, int> GetWordsCounts();
+    }
+}
